requirements.txt
# requirements.txt

# Configuration
python-dotenv
configparser

# Utilities
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Testing
pytest
pytest-mock